Data and code used for Nitrogen emissions
This folder contains all the data extracted from GLEAM2.0 model used to estimate the life-cycle nitrogen use indicators, nitrogen flows and emissions. to use it, please customise the paths in the R code to your computer source. These data cannot be used for any publication without the authorization of the corresponding author.